refactor single-use methods
